Refaktorer: Flytt klasser featurebasert
