+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/MockServer.java
@@ -1,4 +1,4 @@
-package no.nav.tag.sykefravarsstatistikk.api.mock;
+package no.nav.tag.sykefravarsstatistikk.api;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
 import com.github.tomakehurst.wiremock.client.WireMock;
