--- src/main/java/no/nav/tag/sykefravarsstatistikk/api/domain/Identifikator.java
@@ -1,5 +1,0 @@
-package no.nav.tag.sykefravarsstatistikk.api.domain;
-
-public interface Identifikator {
-    String asString();
-}
