+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/domene/InnloggetSelvbetjeningBruker.java
@@ -1,0 +1,39 @@
+package no.nav.tag.sykefravarsstatistikk.api.domene;
+
+import lombok.Data;
+import no.nav.tag.sykefravarsstatistikk.api.altinn.Organisasjon;
+import no.nav.tag.sykefravarsstatistikk.api.tilgangskontroll.TilgangskontrollException;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Objects;
+import java.util.Optional;
+
+@Data
+public class InnloggetSelvbetjeningBruker {
+    private List<Organisasjon> organisasjoner;
+    private Fnr fnr;
+
+    public InnloggetSelvbetjeningBruker(Fnr fnr) {
+        this.fnr = fnr;
+        organisasjoner = new ArrayList<>();
+    }
+
+    public void sjekkTilgang(Orgnr orgnr) {
+        if (!harTilgang(orgnr)) {
+            throw new TilgangskontrollException("Har ikke tilgang til statistikk for denne bedriften.");
+        }
+    }
+
+    protected boolean harTilgang(Orgnr orgnr) {
+        Optional<Organisasjon> organisasjon = organisasjoner
+                .stream()
+                .filter(Objects::nonNull)
+                .filter(
+                        o -> o.getOrganizationNumber().equals(orgnr.getVerdi())
+                )
+                .findFirst();
+
+        return organisasjon.isPresent();
+    }
+}
