+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/domene/statistikk/LandStatistikk.java
@@ -1,4 +1,4 @@
-package no.nav.tag.sykefravarsstatistikk.api.domain.stats;
+package no.nav.tag.sykefravarsstatistikk.api.domene.statistikk;
 
 import lombok.*;
