+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/healthcheck/HealthcheckController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class HealthcheckController {
 
-
     @GetMapping("/internal/healthcheck")
     public String healthcheck() {
         return "ok";
