+++ src/main/java/no/nav/tag/sykefravarsstatistikk/api/statistikk/OpenSykefraværsstatistikkController.java
@@ -1,9 +1,9 @@
-package no.nav.tag.sykefravarsstatistikk.api.controller;
+package no.nav.tag.sykefravarsstatistikk.api.statistikk;
 
 import lombok.extern.slf4j.Slf4j;
 import no.nav.security.oidc.api.Unprotected;
-import no.nav.tag.sykefravarsstatistikk.api.sykefravarprosent.Sykefraværprosent;
-import no.nav.tag.sykefravarsstatistikk.api.sykefravarprosent.SykefraværprosentService;
+import no.nav.tag.sykefravarsstatistikk.api.statistikk.Sykefraværprosent;
+import no.nav.tag.sykefravarsstatistikk.api.statistikk.SykefraværprosentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,12 +15,10 @@ public class OpenSykefraværsstatistikkController {
 
     private final SykefraværprosentService service;
 
-
     @Autowired
     public OpenSykefraværsstatistikkController(SykefraværprosentService service){
         this.service = service;
     }
-
 
     @GetMapping(value = "/sykefravarprosent")
     public Sykefraværprosent sykefraværprosent() {
