+++ src/test/java/no/nav/tag/sykefravarsstatistikk/api/domene/stats/LandStatistikkTest.java
@@ -1,5 +1,6 @@
-package no.nav.tag.sykefravarsstatistikk.api.domain.stats;
+package no.nav.tag.sykefravarsstatistikk.api.domene.stats;
 
+import no.nav.tag.sykefravarsstatistikk.api.domene.statistikk.LandStatistikk;
 import org.junit.Test;
 
 import java.math.BigDecimal;
